refactor: split launcher.py in separate files so it's easier to maintain (mostly into main.py, gui/application.py, core/scanner.py)
