+++ src/gui/fomod_dialog.py
@@ -1,11 +1,6 @@
-#fomod_handler.py
+# src/gui/dialogs.py
 
-#global imports
-import os, zipfile, yaml, shutil, gi
-
-#specific imports
-import xml.etree.ElementTree as ET
-from gi.repository import Gtk, Adw
+from gi.repository import Gtk
 
 class FomodSelectionDialog(Gtk.Dialog):
     def __init__(self, parent, module_name, options):
@@ -17,7 +12,6 @@ class FomodSelectionDialog(Gtk.Dialog):
         content_area = self.get_content_area()
         content_area.set_spacing(15)
         
-        # Manually setting margins because Gtk.Box doesn't support set_margin_all
         content_area.set_margin_top(20)
         content_area.set_margin_bottom(20)
         content_area.set_margin_start(20)
@@ -27,22 +21,17 @@ class FomodSelectionDialog(Gtk.Dialog):
         header.add_css_class("title-2")
         content_area.append(header)
 
-        # 1. Create the ListBox to house our clickable rows
         self.list_box = Gtk.ListBox(css_classes=["boxed-list"])
         self.list_box.set_selection_mode(Gtk.SelectionMode.NONE)
-        
-        # Connect the signal so clicking anywhere in the row toggles the radio
         self.list_box.connect("row-activated", self.on_row_activated)
 
         first_radio = None
         
         for name, desc, source in options:
-            # 2. Create the Radio Button
             radio = Gtk.CheckButton(group=first_radio)
             if not first_radio:
                 first_radio = radio
             
-            # 3. Create the Row Layout
             row_content = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
             row_content.set_margin_start(12)
             row_content.set_margin_end(12)
@@ -64,67 +53,32 @@ class FomodSelectionDialog(Gtk.Dialog):
             row_content.append(radio)
             row_content.append(text_vbox)
 
-            # 4. Wrap everything in a ListBoxRow
             row = Gtk.ListBoxRow()
             row.set_child(row_content)
-            
-            # Custom attribute to link the row back to its radio button
             row.radio_button = radio
             
             self.list_box.append(row)
             self.options_map[radio] = source
 
-        # 5. Add a ScrolledWindow
-        # We add 'vexpand=True' so it grows with the window
         scrolled = Gtk.ScrolledWindow(
             propagate_natural_height=True, 
             vexpand=True, 
             hexpand=True
         )
-        
-        # You can keep max_content_height for the INITIAL size, 
-        # but vexpand will override it once the window is resized.
         scrolled.set_max_content_height(700)
-        
         scrolled.set_child(self.list_box)
         content_area.append(scrolled)
 
-        # 6. Action Buttons
         self.add_button("Install", Gtk.ResponseType.OK)
         self.add_button("Cancel", Gtk.ResponseType.CANCEL)
         self.set_default_response(Gtk.ResponseType.OK)
 
     def on_row_activated(self, list_box, row):
-        """Callback to toggle the radio button when the row is clicked."""
         if hasattr(row, "radio_button"):
             row.radio_button.set_active(True)
 
     def get_selected_source(self):
-        """Returns the source folder associated with the selected radio button."""
         for radio, source in self.options_map.items():
             if radio.get_active():
                 return source
         return None
-
-def parse_fomod_xml(xml_data):
-    """Parses the FOMOD XML and returns a list of (name, description, source_folder)"""
-    try:
-        root = ET.fromstring(xml_data)
-        options = []
-        # Find all plugins (options) within the XML structure
-        for plugin in root.findall(".//plugin"):
-            name = plugin.get('name')
-            desc_node = plugin.find('description')
-            desc = desc_node.text.strip() if desc_node is not None and desc_node.text else "No description provided."
-            
-            folder_node = plugin.find(".//folder")
-            source = folder_node.get('source') if folder_node is not None else None
-            
-            if source:
-                options.append((name, desc, source))
-        
-        module_name = root.findtext('moduleName') or "Unknown Mod"
-        return module_name, options
-    except Exception as e:
-        print(f"Failed to parse FOMOD XML: {e}")
-        return None, []
